SNHU logo now properly shows in center
The SNHU logo was off center when the nav back button appeared. The reason why it was being put off center was because the width was larger than the image. Any extra width was added as padding to the image instead of making it wider. The image padding collided with the back arrow pushing the logo to the left. Using boxfit.contain property in the width attribute determines the correct width to use while keeping the image aspect ratio.